add: rich json
